- osada_write funcionando a todos los niveles (por ahora...) - Modificado el tipo de datos de los offsets y tamaños en las funciones .truncate y write (dejamos de tratarlos como size_t y off_t y los manejamos puramente como ints), debido a bugs en la serialización/recepción de dichas variables - Correción/borrado de algunas funciones secundarias

+++ PokedexServidor/src/libreriaPokedexServidor.h
@@ -154,7 +154,7 @@ void *osada_read(char *unaRuta);
 
 int osada_create(char *ruta);
 
-int osada_write(char *ruta, void *nuevoContenido, size_t sizeAgregado, off_t offset);
+int osada_write(char *ruta, void *nuevoContenido, int sizeAgregado, int offset);
 
 int osada_unlink(char *ruta);
 
@@ -164,6 +164,6 @@ int osada_rmdir(char *ruta);
 
 int osada_rename(char *ruta, char *nuevoNombre);
 
-int osada_truncate(char *ruta, off_t nuevoTamanio);
+int osada_truncate(char *ruta, int nuevoTamanio);
 
 #endif /* LIBRERIAPOKEDEXSERVIDOR_H_ */
